+++ packages/xo-server/src/api/disk.js
@@ -1,5 +1,8 @@
-import { parseSize } from '../utils'
+import pump from 'pump'
+import { format } from 'json-rpc-peer'
 import { unauthorized } from 'xo-common/api-errors'
+
+import { parseSize } from '../utils'
 
 // ===================================================================
 
@@ -11,7 +14,7 @@ export async function create ({ name, size, sr, vm, bootable, position, mode }) 
     await this.checkResourceSetConstraints(resourceSet, this.user.id, [sr.id])
     await this.allocateLimitsInResourceSet({ disk: size }, resourceSet)
   } else if (
-    !await this.hasPermissions(this.user.id, [[sr.id, 'administrate']])
+    !(await this.hasPermissions(this.user.id, [[sr.id, 'administrate']]))
   ) {
     throw unauthorized()
   }
@@ -51,6 +54,83 @@ create.params = {
 create.resolve = {
   vm: ['vm', 'VM', 'administrate'],
   sr: ['sr', 'SR', false],
+}
+
+// -------------------------------------------------------------------
+
+async function handleExportContent (req, res, { xapi, id }) {
+  const stream = await xapi.exportVdiContent(id)
+  req.on('close', () => stream.cancel())
+
+  // Remove the filename as it is already part of the URL.
+  stream.headers['content-disposition'] = 'attachment'
+
+  res.writeHead(
+    stream.statusCode,
+    stream.statusMessage != null ? stream.statusMessage : '',
+    stream.headers
+  )
+  pump(stream, res, error => {
+    if (error != null) {
+      console.warn('disk.exportContent', error)
+    }
+  })
+}
+
+export async function exportContent ({ vdi }) {
+  return {
+    $getFrom: await this.registerHttpRequest(
+      handleExportContent,
+      {
+        id: vdi._xapiId,
+        xapi: this.getXapi(vdi),
+      },
+      {
+        suffix: `/${encodeURIComponent(vdi.name_label)}.vhd`,
+      }
+    ),
+  }
+}
+
+exportContent.description = 'export the content of a VDI'
+exportContent.params = {
+  id: { type: 'string' },
+}
+exportContent.resolve = {
+  vdi: ['id', ['VDI', 'VDI-snapshot'], 'view'],
+}
+
+// -------------------------------------------------------------------
+
+async function handleImportContent (req, res, { xapi, id }) {
+  // Timeout seems to be broken in Node 4.
+  // See https://github.com/nodejs/node/issues/3319
+  req.setTimeout(43200000) // 12 hours
+
+  try {
+    await xapi.importVdiContent(id, req)
+    res.end(format.response(0, true))
+  } catch (e) {
+    res.writeHead(500)
+    res.end(format.error(0, new Error(e.message)))
+  }
+}
+
+export async function importContent ({ vdi }) {
+  return {
+    $sendTo: await this.registerHttpRequest(handleImportContent, {
+      id: vdi._xapiId,
+      xapi: this.getXapi(vdi),
+    }),
+  }
+}
+
+importContent.description = 'import contents into a VDI'
+importContent.params = {
+  id: { type: 'string' },
+}
+importContent.resolve = {
+  vdi: ['id', ['VDI'], 'operate'],
 }
 
 // -------------------------------------------------------------------