+++ packages/xo-web/src/common/dropzone/index.js
@@ -8,16 +8,18 @@ import styles from './index.css'
 @propTypes({
   onDrop: propTypes.func,
   message: propTypes.node,
+  multiple: propTypes.bool,
 })
 export default class Dropzone extends Component {
   render () {
-    const { onDrop, message } = this.props
+    const { onDrop, message, multiple } = this.props
 
     return (
       <ReactDropzone
-        onDrop={onDrop}
-        className={styles.dropzone}
         activeClassName={styles.activeDropzone}
+        className={styles.dropzone}
+        multiple={multiple}
+        onDrop={onDrop}
       >
         <div className={styles.dropzoneText}>{message}</div>
       </ReactDropzone>