+++ packages/xo-web/src/xo-app/settings/remotes/index.js
@@ -1,12 +1,8 @@
 import _, { messages } from 'intl'
 import ActionButton from 'action-button'
 import Component from 'base-component'
-import filter from 'lodash/filter'
 import Icon from 'icon'
-import isEmpty from 'lodash/isEmpty'
-import map from 'lodash/map'
 import React from 'react'
-import some from 'lodash/some'
 import SortedTable from 'sorted-table'
 import StateButton from 'state-button'
 import Tooltip from 'tooltip'
@@ -14,9 +10,10 @@ import { addSubscriptions } from 'utils'
 import { alert, confirm } from 'modal'
 import { error } from 'notification'
 import { format, parse } from 'xo-remote-parser'
+import { groupBy, map, isEmpty, some } from 'lodash'
 import { injectIntl } from 'react-intl'
-import { Number, Password, Text } from 'editable'
 import { Number as InputNumber } from 'form'
+import { Number, Password, Text } from 'editable'
 
 import {
   createRemote,
@@ -34,25 +31,21 @@ const remoteTypes = {
   nfs: 'remoteTypeNfs',
   smb: 'remoteTypeSmb',
 }
-const _changeUrlElement = (remote, value, element) =>
+const _changeUrlElement = (value, { remote, element }) =>
   editRemote(remote, {
     url: format({ ...remote, [element]: value === null ? undefined : value }),
   })
 const _showError = remote => alert(_('remoteConnectionFailed'), remote.error)
+const _editRemote = (name, { remote }) => editRemote(remote, { name })
 const COLUMN_NAME = {
-  component: @injectIntl
-  class RemoteName extends Component {
-    render () {
-      const { item: remote, intl } = this.props
-      return (
-        <Text
-          onChange={name => editRemote(remote, { name })}
-          placeholder={intl.formatMessage(messages.remoteMyNamePlaceHolder)}
-          value={remote.name}
-        />
-      )
-    }
-  },
+  itemRenderer: (remote, { formatMessage }) => (
+    <Text
+      data-remote={remote}
+      onChange={_editRemote}
+      placeholder={formatMessage(messages.remoteMyNamePlaceHolder)}
+      value={remote.name}
+    />
+  ),
   name: _('remoteName'),
   sortCriteria: 'name',
 }
@@ -88,21 +81,15 @@ const COLUMN_STATE = {
 const COLUMNS_LOCAL_REMOTE = [
   COLUMN_NAME,
   {
-    component: @injectIntl
-    class LocalRemotePath extends Component {
-      render () {
-        const { item: remote, intl } = this.props
-        return (
-          <Text
-            onChange={v => _changeUrlElement(remote, v, 'path')}
-            placeholder={intl.formatMessage(
-              messages.remoteLocalPlaceHolderPath
-            )}
-            value={remote.path}
-          />
-        )
-      }
-    },
+    itemRenderer: (remote, { formatMessage }) => (
+      <Text
+        data-element='path'
+        data-remote={remote}
+        onChange={_changeUrlElement}
+        placeholder={formatMessage(messages.remoteLocalPlaceHolderPath)}
+        value={remote.path}
+      />
+    ),
     name: _('remotePath'),
   },
   COLUMN_STATE,
@@ -110,41 +97,36 @@ const COLUMNS_LOCAL_REMOTE = [
 const COLUMNS_NFS_REMOTE = [
   COLUMN_NAME,
   {
-    component: @injectIntl
-    class NfsRemoteInfo extends Component {
-      render () {
-        const { item: remote, intl } = this.props
-        return (
-          <span>
-            <strong className='text-info'>\\</strong>
-            <Text
-              onChange={v => _changeUrlElement(remote, v, 'host')}
-              placeholder={intl.formatMessage(
-                messages.remoteNfsPlaceHolderHost
-              )}
-              value={remote.host}
-            />
-            :
-            <Number
-              nullable
-              onChange={v => _changeUrlElement(remote, v, 'port')}
-              placeholder={intl.formatMessage(
-                messages.remoteNfsPlaceHolderPort
-              )}
-              value={remote.port || ''}
-            />
-            :
-            <Text
-              onChange={v => _changeUrlElement(remote, v, 'path')}
-              placeholder={intl.formatMessage(
-                messages.remoteNfsPlaceHolderPath
-              )}
-              value={remote.path}
-            />
-          </span>
-        )
-      }
-    },
+    itemRenderer: (remote, { formatMessage }) => (
+      <span>
+        <strong className='text-info'>\\</strong>
+        <Text
+          data-element='host'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          placeholder={formatMessage(messages.remoteNfsPlaceHolderHost)}
+          value={remote.host}
+        />
+        :
+        <Number
+          data-element='port'
+          data-remote={remote}
+          nullable
+          onChange={_changeUrlElement}
+          placeholder={formatMessage(messages.remoteNfsPlaceHolderPort)}
+          value={remote.port || ''}
+        />
+        :
+        <Text
+          data-element='path'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          placeholder={formatMessage(messages.remoteNfsPlaceHolderPath)}
+          value={remote.path}
+        />
+      </span>
+    ),
+
     name: _('remoteDevice'),
   },
   COLUMN_STATE,
@@ -152,62 +134,56 @@ const COLUMNS_NFS_REMOTE = [
 const COLUMNS_SMB_REMOTE = [
   COLUMN_NAME,
   {
-    component: @injectIntl
-    class SmbRemoteInfo extends Component {
-      render () {
-        const { item: remote, intl } = this.props
-        return (
-          <span>
-            <strong className='text-info'>\\</strong>
-            <Text
-              value={remote.host}
-              onChange={v => _changeUrlElement(remote, v, 'host')}
-            />
-            <strong className='text-info'>\</strong>
-            <span>
-              <Text
-                onChange={v => _changeUrlElement(remote, v, 'path')}
-                placeholder={intl.formatMessage(
-                  messages.remoteSmbPlaceHolderRemotePath
-                )}
-                value={remote.path}
-              />
-            </span>
-          </span>
-        )
-      }
-    },
+    itemRenderer: (remote, { formatMessage }) => (
+      <span>
+        <strong className='text-info'>\\</strong>
+        <Text
+          data-element='host'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          value={remote.host}
+        />
+        <strong className='text-info'>\</strong>
+        <span>
+          <Text
+            data-element='path'
+            data-remote={remote}
+            onChange={_changeUrlElement}
+            placeholder={formatMessage(messages.remoteSmbPlaceHolderRemotePath)}
+            value={remote.path}
+          />
+        </span>
+      </span>
+    ),
     name: _('remoteShare'),
   },
   COLUMN_STATE,
   {
-    component: @injectIntl
-    class SmbRemoteAuthInfo extends Component {
-      render () {
-        const { item: remote, intl } = this.props
-        return (
-          <span>
-            <Text
-              value={remote.username}
-              onChange={v => _changeUrlElement(remote, v, 'username')}
-            />
-            :
-            <Password
-              value=''
-              onChange={v => _changeUrlElement(remote, v, 'password')}
-              placeholder={intl.formatMessage(
-                messages.remotePlaceHolderPassword
-              )}
-            />
-            @
-            <Text
-              value={remote.domain}
-              onChange={v => _changeUrlElement(remote, v, 'domain')}
-            />
-          </span>
-        )
-      }
-    },
+    itemRenderer: (remote, { formatMessage }) => (
+      <span>
+        <Text
+          data-element='username'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          value={remote.username}
+        />
+        :
+        <Password
+          data-element='password'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          placeholder={formatMessage(messages.remotePlaceHolderPassword)}
+          value=''
+        />
+        @
+        <Text
+          data-element='domain'
+          data-remote={remote}
+          onChange={_changeUrlElement}
+          value={remote.domain}
+        />
+      </span>
+    ),
     name: _('remoteAuth'),
   },
 ]
@@ -268,22 +244,22 @@ const FILTERS = {
 
 @addSubscriptions({
   remotes: cb =>
-    subscribeRemotes(rawRemotes => {
-      rawRemotes = map(rawRemotes, remote => {
-        try {
-          return {
-            ...remote,
-            ...parse(remote.url),
-          }
-        } catch (err) {
-          console.error('Remote parsing error:', remote, '\n', err)
-        }
-      }).filter(r => r !== undefined)
-      const remotes = {}
-      for (const remoteType in remoteTypes) {
-        remotes[remoteType] = filter(rawRemotes, r => r.type === remoteType)
-      }
-      cb(remotes)
+    subscribeRemotes(remotes => {
+      cb(
+        groupBy(
+          map(remotes, remote => {
+            try {
+              return {
+                ...remote,
+                ...parse(remote.url),
+              }
+            } catch (err) {
+              console.error('Remote parsing error:', remote, '\n', err)
+            }
+          }).filter(r => r !== undefined),
+          'type'
+        )
+      )
     }),
 })
 @injectIntl
@@ -360,7 +336,10 @@ export default class Remotes extends Component {
   }
 
   render () {
-    const { remotes = {} } = this.props
+    const {
+      intl: { formatMessage },
+      remotes = {},
+    } = this.props
     const {
       domain,
       host,
@@ -380,6 +359,7 @@ export default class Remotes extends Component {
             <SortedTable
               collection={remotes.file}
               columns={COLUMNS_LOCAL_REMOTE}
+              data-formatMessage={formatMessage}
               filters={FILTERS}
               groupedActions={GROUPED_ACTIONS}
               individualActions={INDIVIDUAL_ACTIONS}
@@ -394,6 +374,7 @@ export default class Remotes extends Component {
             <SortedTable
               collection={remotes.nfs}
               columns={COLUMNS_NFS_REMOTE}
+              data-formatMessage={formatMessage}
               filters={FILTERS}
               groupedActions={GROUPED_ACTIONS}
               individualActions={INDIVIDUAL_ACTIONS}
@@ -408,6 +389,7 @@ export default class Remotes extends Component {
             <SortedTable
               collection={remotes.smb}
               columns={COLUMNS_SMB_REMOTE}
+              data-formatMessage={formatMessage}
               filters={FILTERS}
               groupedActions={GROUPED_ACTIONS}
               individualActions={INDIVIDUAL_ACTIONS}
@@ -441,9 +423,7 @@ export default class Remotes extends Component {
             <input
               className='form-control'
               onChange={this.linkState('name')}
-              placeholder={this.props.intl.formatMessage(
-                messages.remoteMyNamePlaceHolder
-              )}
+              placeholder={formatMessage(messages.remoteMyNamePlaceHolder)}
               required
               type='text'
               value={name}
@@ -457,7 +437,7 @@ export default class Remotes extends Component {
                   className='form-control'
                   onChange={this.linkState('path')}
                   pattern='^(([^/]+)+(/[^/]+)*)?$'
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteLocalPlaceHolderPath
                   )}
                   type='text'
@@ -472,9 +452,7 @@ export default class Remotes extends Component {
                 <input
                   className='form-control'
                   onChange={this.linkState('host')}
-                  placeholder={this.props.intl.formatMessage(
-                    messages.remoteNfsPlaceHolderHost
-                  )}
+                  placeholder={formatMessage(messages.remoteNfsPlaceHolderHost)}
                   type='text'
                   value={host}
                   required
@@ -482,9 +460,7 @@ export default class Remotes extends Component {
                 <br />
                 <InputNumber
                   onChange={this.linkState('port')}
-                  placeholder={this.props.intl.formatMessage(
-                    messages.remoteNfsPlaceHolderPort
-                  )}
+                  placeholder={formatMessage(messages.remoteNfsPlaceHolderPort)}
                   value={port}
                 />
               </div>
@@ -494,9 +470,7 @@ export default class Remotes extends Component {
                   className='form-control'
                   onChange={this.linkState('path')}
                   pattern='^(([^/]+)+(/[^/]+)*)?$'
-                  placeholder={this.props.intl.formatMessage(
-                    messages.remoteNfsPlaceHolderPath
-                  )}
+                  placeholder={formatMessage(messages.remoteNfsPlaceHolderPath)}
                   type='text'
                   value={path}
                 />
@@ -511,7 +485,7 @@ export default class Remotes extends Component {
                   className='form-control'
                   onChange={this.linkState('host')}
                   pattern='^([^\\/]+)\\([^\\/]+)$'
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteSmbPlaceHolderAddressShare
                   )}
                   type='text'
@@ -523,7 +497,7 @@ export default class Remotes extends Component {
                   className='form-control'
                   onChange={this.linkState('path')}
                   pattern='^(([^\\/]+)+(\\[^\\/]+)*)?$'
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteSmbPlaceHolderRemotePath
                   )}
                   type='text'
@@ -534,7 +508,7 @@ export default class Remotes extends Component {
                 <input
                   className='form-control'
                   onChange={this.linkState('username')}
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteSmbPlaceHolderUsername
                   )}
                   type='text'
@@ -545,7 +519,7 @@ export default class Remotes extends Component {
                 <input
                   className='form-control'
                   onChange={this.linkState('password')}
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteSmbPlaceHolderPassword
                   )}
                   type='text'
@@ -556,7 +530,7 @@ export default class Remotes extends Component {
                 <input
                   className='form-control'
                   onChange={this.linkState('domain')}
-                  placeholder={this.props.intl.formatMessage(
+                  placeholder={formatMessage(
                     messages.remoteSmbPlaceHolderDomain
                   )}
                   required